modul-08/requirements.txt
openai
python-dotenv
ipython
opencv-python
moviepy
azure-cognitiveservices-speech